feat: Implement question answering
Well, that escalated quickly. Redux sucks some of the joy out of React coding, but it gets the job done.
I'm not happy with the way I've done things here. It wouldn't be easy to follow for another dev (or myself in a few weeks).
TODO: Look at Question.js and pull some functionality out into their own components.