fix: Modify the navigatin/routing for the question detail page
Reviewer's comments:
"The details of the poll should be available at questions/:question_id. Currently, it is available at question-detail/:question_id
If a poll does not exist, the application should show a 404 page after the user has signed in. Example- If the user visits http://localhost:3000/question-detail/invalid from the address bar, after login, the user should be shown a 404 page. Currently, the user is shown the 404 page without login."